requirements.txt
asgiref==3.8.1
astroid==3.3.5
autopep8==2.3.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
phonenumbers==8.13.50
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg==3.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.0
pylint==3.3.1
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
redis==5.2.0
requests==2.32.3
sqlparse==0.5.2
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
